Refactor flow fetches into bundle service
